requirements.txt
absl-py==2.1.0
appdirs==1.4.4
astunparse==1.6.3
atomicwrites==1.4.0
bs4==0.0.2
distlib==0.3.9
django-grappelli==4.0.2
django-jazzmin==3.0.1
django_db==0.0.7
djangorestframework==3.15.2
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
gunicorn==23.0.0
jsonpointer==2.1
keras==3.8.0
libclang==18.1.1
matplotlib==3.9.2
ml-dtypes==0.3.2
namex==0.0.8
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
protobuf==4.25.3
pyasn1-modules==0.2.8
pyls-spyder==0.4.0
PyMySQL==1.1.1
PyQt5==5.15.10
PyQtWebEngine==5.15.6
scikit-learn==1.6.1
setuptools==75.1.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
termcolor==2.5.0
torch==2.2.2
torchvision==0.17.2
virtualenv==20.29.1
wheel==0.44.0
xgboost==2.1.4